# Enable async profiler