In exploring [FLINK-35215](https://issues.apache.org/jira/browse/FLINK-35215) being able to generate flame graphs from each benchmark run makes comparing results a lot easier.

However specifying the configuration of the profiler using the exec plugin is problematic as we need to conditionally add an argument based on the value of a property. So instead of duplicating the existing profiles to add profiler when the new profiles are activated delegate the execution to a shell script which can handle the optional parameters.