@SamBarker 2024-05-27
I think the `benchmark-java17` profile could potentially be simplified to specify a property with just the additional JVM_ARGS (possibly changing `-j` to be single arg inline with `-p`). 

However I wanted to gauge interest in this change before I went any further with potential refactoring. 

I also think it would be very helpful to capture the flamegraphs as part of the information included in `flink-speed` but I have no idea about the implications of that or the required setup for the build server so haven't perused that option as yet.