@SamBarker 2024-05-27
> AFAIU, by default profiling is not turned on. One has to activate enable-async-profiler profile?

Yes that's almost right. The profile is enabled automatically by specifying the path the profiler shared library (as opposed to running `-P enable-async-profiler`) Async profiler needs downloaded manually and the path is platform specific I didn't want to tackle automation of that unless there was some interest in the idea.