@SamBarker 2024-05-29
@Zakelly any news on the jenkins env? I'm largely hoping its a no-op there right now.

@pnowojski readme and executable bit set. Is there anything else your looking for?