@Zakelly 2024-05-29
And I find the benchmark runs extremely slow with the profiler on. The ETA is up to 8 days for full set of our daily run. So I guess this is best only used for single tests that are triggered manually, right?