@SamBarker 2024-05-29
> And I find the benchmark runs extremely slow with the profiler on. The ETA is up to 8 days for full set of our daily run. So I guess this is best only used for single tests that are triggered manually, right?

Yikes thats a bit sad. Running a single test (`SerializationFrameworkMiniBenchmarks.serializerHeavyString`) locally I get 10 mins estimate with or without the profiler and a runtime that seems comparable. If I widen the set of benchmarks I get the same ETA with or without the profiler. 

So I'm slightly suspicious that the excludes or default includes behaviour is being changed but without knowing what command line your seeing that from its hard to know.